refactor: rename fromFixture to fromGenericObjects
