+++ apis/enigma-mocker/src/__tests__/from-generic-objects.spec.js
@@ -1,4 +1,4 @@
-import createEnigmaMocker from '../from-fixture';
+import createEnigmaMocker from '../from-generic-objects';
 
 const fixture = {
   getLayout: {
